Deploy simple CR7 RTOS branch
This branch is used by eLinux documentation explaining how to deploy
Zephyr on R-Car Gen3 boards.
This tutorial needs to provide an easy way to patch the BL2 bootloader
in order for the CR7 core to receive interrupts from the GIC controller.
This branch is then used only to patch BL2 for BSPs v5.5.0 and later.

Signed-off-by: Aymeric Aillet <aymeric.aillet@iot.bzh>