[DRAFT] Add support for falcon/spider to remoteproc
Backport upstream patch from:
 * https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/log/?qt=author&q=massot
 The patch from upstream has been fix to be compatible with the
spider/falcon RCAR BSP.

Those patches need to be check

The spider need to have a support for the cortex R7 (the name need to be check)(no section "reserved-memory" found)

The falcon need to have correct clock value for mfis.

The mailbox need to be check(The new Kconfig is correct?).

Remove useless patch

Signed-off-by: Ronan Le Martret <ronan.lemartret@iot.bzh>